chore: getting multi environments to work
