+++ ksim/model/formulations.py
@@ -4,6 +4,8 @@ from abc import ABC, abstractmethod
 from typing import Tuple
 
 import flax.linen as nn
+import jax
+import jax.numpy as jnp
 from flax.core import FrozenDict
 from jaxtyping import Array, PRNGKeyArray
 
@@ -32,10 +34,35 @@ class ActionModel(nn.Module, ABC):
 class GaussianActionModel(ActionModel, ABC):
     """Gaussian action model."""
 
-    @abstractmethod
-    def get_variance(self) -> Array:
-        """Get the variance of the action model."""
-        ...
+    init_log_std: float
+    num_outputs: int
+
+    def setup(self) -> None:
+        self.log_std = self.param(
+            "log_std", nn.initializers.constant(self.init_log_std), (self.num_outputs,)
+        )
+
+    def calc_log_prob(self, prediction: Array, action: Array) -> Array:
+        mean = prediction
+        std = jnp.exp(self.log_std)
+
+        log_prob = (
+            -0.5 * jnp.square((action - mean) / std) - jnp.log(std) - 0.5 * jnp.log(2 * jnp.pi)
+        )  # (...batch_dims..., action_dim)
+        # if we assume indep gaussians, can just sum over action dim in log space for log prob
+        return jnp.sum(log_prob, axis=-1)
+
+    def sample_and_log_prob(
+        self, obs: FrozenDict[str, Array], cmd: FrozenDict[str, Array], rng: PRNGKeyArray
+    ) -> Tuple[Array, Array]:
+        mean = self(obs, cmd)
+        std = jnp.exp(self.log_std)
+
+        noise = jax.random.normal(rng, mean.shape)
+        action = mean + noise * std
+        log_prob = self.calc_log_prob(mean, action)
+
+        return action, log_prob
 
 
 class CategoricalActionModel(ActionModel, ABC):
